Renames the shapeSizesSlide to Slice, which is what it is :)
